- The interface to generated code is now in a struct callback, to avoid conflicting names when we load dynamic libraries (FMU's) - Fixed lots of clang warnings
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@18260 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/math-support/ddassl.c
@@ -3765,7 +3765,7 @@ L100:
       INDENT(LOG_JAC);
       for(k=0;k<*neq;k++)
       {
-        INFO2(LOG_JAC, "[%ld] %g", k+1, y[k+1]);
+        INFO2(LOG_JAC, "[%ld] %g", (long) k+1, y[k+1]);
       }
       RELEASE(LOG_JAC);
 
@@ -3845,7 +3845,7 @@ L200:
       INDENT(LOG_JAC);
       for(k=0;k<*neq;k++)
       {
-        INFO2(LOG_JAC, "[%ld] %g", k+1, y[k+1]);
+        INFO2(LOG_JAC, "[%ld] %g", (long) k+1, y[k+1]);
       }
       RELEASE(LOG_JAC);
